pyproject.toml
[project]
name = "helf"
version = "0.1.0"
description = "Simple workout tracking app with CSV-based storage"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "nicegui>=2.10.0",
    "plotly>=5.24.0",
    "paho-mqtt>=2.0.0",
    "numpy>=2.3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
